pyproject.toml
[project]
name = "pet-app"
version = "0.1.0"
description = "Simple FastAPI app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.4",
    "bcrypt>=4.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "passlib>=1.7.4",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "python-dotenv==1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy==2.0.42",
    "typing-extensions==4.14.1",
    "uvicorn==0.35.0",
]
